# api/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: MetalDocs API
  version: 1.0.0
  description: >
    API inicial do MetalDocs (v1). Escopo atual sem integracao com IA.
servers:
  - url: http://192.168.0.3:8080/api/v1
paths:
...
  /auth/login:
    post:
      summary: Autentica usuario local e cria sessao por cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Sessao criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          description: Credenciais invalidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /auth/me:
    get:
      summary: Retorna o usuario autenticado atual
      responses:
        '200':
          description: Usuario atual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /auth/change-password:
    post:
      summary: Troca a senha do usuario autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Senha alterada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '400':
          description: Requisicao invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
components:
...
  schemas:
    AuthLoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
        password:
          type: string
    CurrentUser:
      type: object
      required: [userId, username, displayName, mustChangePassword, roles]
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        mustChangePassword:
          type: boolean
        roles:
          type: array
          items:
            type: string
            enum: [admin, editor, reviewer, viewer]
    AuthLoginResponse:
      type: object
      required: [user, expiresAt]
      properties:
        user:
          $ref: '#/components/schemas/CurrentUser'
        expiresAt:
          type: string
          format: date-time
    ChangePasswordRequest:
      type: object
      required: [newPassword]
      properties:
        currentPassword:
          type: string
          description: Obrigatoria para troca de senha em sessao normal; opcional no fluxo de primeiro acesso com troca obrigatoria.
        newPassword:
          type: string
    ChangePasswordResponse:
      type: object
      required: [changed, user]
      properties:
        changed:
          type: boolean
        user:
          $ref: '#/components/schemas/CurrentUser'
...
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message, details, trace_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string